src/Onkyo Network Remote.indigoPlugin/Contents/Server Plugin/requirements.txt
onkyo-eiscp==1.2.7